requirements.txt
streamlit
python-dotenv
openai
Flask
streamlit_extras